display pizzas complete
